search/microservices/search/requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
grpcio==1.30.0
grpcio-tools==1.30.0
idna==2.10
multidict==4.7.6
protobuf==3.12.2
pyparsing==2.4.7
six==1.15.0
uvloop==0.14.0
yarl==1.4.2
